+++ segtypes/n64/group.py
@@ -1,43 +1,11 @@
+from segtypes.n64.data import N64SegData
+from segtypes.n64.bss import N64SegBss
+from segtypes.n64.rodata import N64SegRodata
 from typing import List, Dict, Optional
 import sys
 from segtypes.n64.segment import N64Segment
 from segtypes.segment import RomAddr, Segment
 
-    # def get_out_subdir(self) -> Path:
-    #     if self.type.startswith("."):
-    #         if self.c_sibling:
-    #             return self.c_sibling.get_out_subdir()
-    #         else:
-    #             return options.get_src_path()
-    #     elif self.type in ["c"]:
-    #         return options.get_src_path()
-    #     elif self.type in ["asm", "hasm"]:
-    #         return options.get_asm_path()
-    #     elif self.type in ["rodata"]:
-    #         return options.get_asm_path() / "data"
-
-    #     return options.get_asset_path()
-
-    # def get_ld_obj_type(self):
-    #     if self.type in ["c", "asm", "hasm"]:
-    #         return ".text"
-    #     elif self.type in [".rodata", "rodata"]:
-    #         return ".rodata"
-
-    # def get_ext(self):
-    #     if self.type.startswith("."):
-    #         if self.c_sibling:
-    #             return self.c_sibling.get_ext()
-    #         else:
-    #             return "c"
-    #     elif self.type in ["c"]:
-    #         return "c"
-    #     elif self.type in ["asm", "hasm"]:
-    #         return "s"
-    #     elif self.type in ["rodata"]:
-    #         return self.type + ".s"
-    #     return self.type
-        
 class N64SegGroup(N64Segment):
     def __init__(self, segment, rom_start, rom_end):
         super().__init__(segment, rom_start, rom_end)
@@ -73,22 +41,28 @@ class N64SegGroup(N64Segment):
                 sys.exit(1)
 
             segment: Segment = segment_class(subsection_yaml, start, end)
-            segment.c_sibling = base_segments.get(segment.name, None)
+            segment.sibling = base_segments.get(segment.name, None)
             segment.parent = self
 
-            if self.rodata_vram_start == -1 and "rodata" in typ:
+            # TODO: assumes section order - generalize and stuff
+            if self.data_vram_start == None and isinstance(segment, N64SegData):
+                self.data_vram_start = segment.vram_start
+            if self.rodata_vram_start == None and isinstance(segment, N64SegRodata):
+                self.data_vram_end = segment.vram_start
                 self.rodata_vram_start = segment.vram_start
-            if self.rodata_vram_end == -1 and "bss" in typ:
+            if self.rodata_vram_end == None and isinstance(segment, N64SegBss):
                 self.rodata_vram_end = segment.vram_start
+                self.bss_vram_start = segment.vram_start
 
             ret.append(segment)
 
+            # todo change
             if typ in ["c", "asm", "hasm"]:
                 base_segments[segment.name] = segment
 
             prev_start = start
 
-        if self.rodata_vram_start != -1 and self.rodata_vram_end == -1:
+        if self.rodata_vram_start != None and self.rodata_vram_end == None:
             assert self.vram_end is not None
             self.rodata_vram_end = self.vram_end
 
@@ -97,7 +71,7 @@ class N64SegGroup(N64Segment):
     def get_linker_entries(self):
         return [sub.get_linker_entry() for sub in self.subsegments]
 
-    def get_subsection_for_ram(self, addr):
+    def get_subsegment_for_ram(self, addr) -> Optional[Segment]:
         for sub in self.subsegments:
             if sub.contains_vram(addr):
                 return sub
